Double buffer the player
Gstreamer buffers try to be time based. They try to guess when enough data is buffered to start playback based on the input stream bitrate and codecs. For network streaming it does not work so well. This adds an additional buffer that is size based. We then ignore the primary time based buffer and only listen to the secondary size based buffer. This results in more reliable and expected buffer behaviour(the buffer doesn't flutter and debugging messages that make more sense).