Tame buffering
This changes how we handle buffer messages. Since the percent we get in the buffering handler is pretty much useless this uses a buffer query to get a bool representing the buffering state.

It also changes the timeout of the httpsoupsrc element to give the user plenty of time to reconnect in case of network drops.